Stupid bug has been here the whole time
Basically, (1 << bundleOffset) is using std::uint32_t for the "1" instead of size_t for the "1"

Therefore, sets like 10576 (part of bundle 34) were saying "Yes, I am part of bundle 34", so the initializer was OR-ing it with (1 << 34)... which is apparantly 1 << 2, , or bundle 2, instead of 34.

+++ main.cu
@@ -114,7 +114,7 @@ void initializeSetBundles(size_t numBundles, size_t numSeries, size_t** bundleDa
             for(size_t bundleOffset = 0; bundleOffset < (sizeof(size_t)*8); bundleOffset++){
                 size_t bundleNumToCheck = (sizeof(size_t) * i) + bundleOffset;
                 if(bundleContainsSet(setNum, bundleNumToCheck, numBundles, numSeries, bundleData)){
-                    setBundlesValue = setBundlesValue | (1 << bundleOffset);
+                    setBundlesValue = setBundlesValue | (((size_t)1) << bundleOffset);
                 }
             }
             host_setBundles[setBundlesIdx] = setBundlesValue;
